requirements.txt
opencv-python
numpy
scikit-image
matplotlib
google-cloud-speech
